+++ Content.Shared/SimpleStation14/Holograms/HologramProjectorComponent.cs
@@ -3,4 +3,5 @@ namespace Content.Shared.SimpleStation14.Hologram;
 [RegisterComponent]
 public sealed class HologramProjectorComponent : Component
 {
+
 }
