--- Content.Shared/SimpleStation14/Holograms/HolopodComponent.cs
@@ -1,6 +1,0 @@
-namespace Content.Server.SimpleStation14.Hologram;
-
-[RegisterComponent]
-public sealed class HologramPodComponent : Component
-{
-}
