+++ Resources/Prototypes/SimpleStation14/Entities/Mobs/Player/hologram.yml
@@ -139,63 +139,104 @@
     attributes:
       proper: true
   - type: Hologram
+  - type: AnimationPlayer
+  - type: Mind
+    showExamineInfo: true
+  - type: Inventory
+  - type: Actions
+  - type: Eye
 
 - type: entity
   save: false
   name: Urist McLight
+  suffix: Projected
   parent: PlayerHologramBase
-  id: MobHologram
+  id: MobHologramProjected
   components:
-  # - type: CombatMode
-  #   disarm: null
+  - type: Hologram
+    # holoData:
+    #   Type: HoloType.Projected
+    #   IsHardLight: false
   - type: Stealth
-    lastVisibility: 0.83
+    lastVisibility: 0.80
   - type: InteractionPopup
     successChance: 1
     interactSuccessString: hugging-success-hologram
     # interactSuccessSound: /Audio/Effects/thudswoosh.ogg
     messagePerceivedByOthers: hugging-success-hologram-others
-  - type: Mind
-    showExamineInfo: true
-  - type: Input
-    context: "human"
-  - type: MobMover
-  - type: InputMover
-  - type: Inventory
-  - type: Alerts
-  - type: Actions
-  - type: Eye
-  # - type: CameraRecoil
-  - type: Examiner
   - type: Faction
     factions:
     - NanoTrasen
-  - type: IntrinsicRadioReceiver
-    channels:
-    - Common
-    - Engineering
-    - Medical
-    - Science
-    - Service
-    - Binary
-  - type: IntrinsicRadioTransmitter
-    channels:
-    - Common
-    - Engineering
-    - Medical
-    - Science
-    - Service
-    - Binary
-  - type: ActiveRadio
-    channels:
-    - Common
-    - Engineering
-    - Medical
-    - Science
-    - Service
-    - Binary
   - type: PointLight
-    radius: 1.3
-    softness: 1
+    radius: 1.0
+    softness: 1.4
     color: "#00FFFF"
     energy: 3
+
+- type: entity
+  save: false
+  name: Urist McLight
+  suffix: Lightbee
+  parent: PlayerHologramBase
+  id: MobHologramLightbee
+  components:
+  - type: Hologram
+    # holoData: !HoloData
+    #   Type: Lightbee
+    #   IsHardLight: false
+  - type: Stealth
+    lastVisibility: 0.85
+  - type: InteractionPopup
+    successChance: 1
+    interactSuccessString: hugging-success-hologram
+    # interactSuccessSound: /Audio/Effects/thudswoosh.ogg
+    messagePerceivedByOthers: hugging-success-hologram-others
+  - type: CameraRecoil
+  - type: Faction
+    factions:
+    - NanoTrasen
+  - type: PointLight
+    radius: 1.5
+    softness: 0.8
+    color: "#00FFFF"
+    energy: 4
+
+- type: entity
+  save: false
+  name: Urist McLight
+  suffix: Hardlight
+  parent: PlayerHologramBase
+  id: MobHologramHardlight
+  components:
+  - type: Hologram
+    # holoData:
+    #   Type: HoloType.Lightbee
+    #   IsHardLight: true
+  - type: DamageOnHighSpeedImpact
+    damage:
+      types:
+        Blunt: 5
+    soundHit:
+      path: /Audio/Effects/hit_kick.ogg
+  - type: Damageable
+    damageContainer: Inorganic
+  - type: Pullable
+  - type: Puller
+  - type: CombatMode
+    disarm: null
+  - type: Stealth
+    lastVisibility: 0.90
+  - type: InteractionPopup
+    successChance: 1
+    interactSuccessString: hugging-success-hologram
+    # interactSuccessSound: /Audio/Effects/thudswoosh.ogg
+    messagePerceivedByOthers: hugging-success-hologram-others
+  - type: CameraRecoil
+  - type: Faction
+    factions:
+    - NanoTrasen
+  - type: PointLight
+    radius: 0.8
+    softness: 0.4
+    color: "#00FFFF"
+    energy: 5
