+++ Resources/Prototypes/SimpleStation14/tags.yml
@@ -7,4 +7,6 @@
 - type: Tag
   id: Hardlight
 - type: Tag
+  id: Light
+- type: Tag
   id: HoloDisk
